requirements.txt
# Core
gunicorn>=19.8.1
dash>=1.16.3
dash-bootstrap-components>=0.10.7
# Additional
colorlover>=0.2.1
numpy>=1.16.2
pandas>=0.24.2
scikit-learn>=0.20.3
scipy>=1.2.1
dash-table>=4.10.1
plotly-express>=0.4.1
sqlalchemy>=1.3.20
psycopg2>=2.8.6
dash-core-components>=1.12.1
dash-html-components>=1.1.1
DateTime>=4.3
pyodbc>=4.0.30
